% Auto-generated: do not edit by hand
\name{dashboard}

\alias{dashboard}

\title{Dashboard component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashboard(children=NULL, id=NULL, label=NULL, value=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Children.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{value}{Character. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

