% Auto-generated: do not edit by hand
\name{menu}

\alias{menu}

\title{Menu component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
menu(id=NULL, dashboards=NULL, isOpen=NULL, toggle=NULL)
}

\arguments{
\item{id}{The ID used to identify this component in Dash callbacks.}

\item{dashboards}{The list of dashboards displayed in the sidemenu.}

\item{isOpen}{True if menu is open.}

\item{toggle}{Callback to toggle open state.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

