% Auto-generated: do not edit by hand
\name{widget}

\alias{widget}

\title{Widget component}

\description{
Main is the part of the dashboard that displays all data.
}

\usage{
widget(children=NULL, id=NULL, value=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Children.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{value}{Character. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

